requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
arrow==1.2.3
    # via isoduration
attrs==23.1.0
    # via jsonschema
boto3==1.26.123
    # via neptune
botocore==1.29.123
    # via
    #   boto3
    #   s3transfer
bravado==11.0.3
    # via neptune
bravado-core==5.17.1
    # via bravado
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via neptune
flwr==1.3.0
    # via -r requirements.in
fqdn==1.5.1
    # via jsonschema
future==0.18.3
    # via neptune
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via neptune
grpcio==1.51.3
    # via flwr
idna==3.4
    # via
    #   jsonschema
    #   requests
isoduration==20.11.0
    # via jsonschema
iterators==0.0.2
    # via flwr
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpointer==2.3
    # via jsonschema
jsonref==1.1.0
    # via bravado-core
jsonschema[format]==4.17.3
    # via
    #   bravado-core
    #   swagger-spec-validator
monotonic==1.6
    # via bravado
msgpack==1.0.5
    # via
    #   bravado
    #   bravado-core
neptune==1.1.1
    # via -r requirements.in
numpy==1.24.2
    # via
    #   flwr
    #   pandas
    #   torchvision
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
oauthlib==3.2.2
    # via
    #   neptune
    #   requests-oauthlib
packaging==23.1
    # via neptune
pandas==2.0.1
    # via neptune
pillow==9.4.0
    # via
    #   neptune
    #   torchvision
protobuf==3.20.3
    # via flwr
psutil==5.9.5
    # via neptune
pyjwt==2.4.0
    # via neptune
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   bravado
    #   bravado-core
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via
    #   bravado-core
    #   pandas
pyyaml==6.0
    # via
    #   bravado
    #   bravado-core
    #   swagger-spec-validator
requests==2.28.2
    # via
    #   bravado
    #   bravado-core
    #   neptune
    #   requests-oauthlib
    #   torchvision
requests-oauthlib==1.3.1
    # via neptune
rfc3339-validator==0.1.4
    # via jsonschema
rfc3987==1.3.8
    # via jsonschema
s3transfer==0.6.0
    # via boto3
simplejson==3.19.1
    # via
    #   bravado
    #   bravado-core
six==1.16.0
    # via
    #   bravado
    #   bravado-core
    #   neptune
    #   python-dateutil
    #   rfc3339-validator
smmap==5.0.0
    # via gitdb
swagger-spec-validator==3.0.3
    # via
    #   bravado-core
    #   neptune
torch==1.13.1
    # via
    #   -r requirements.in
    #   torchaudio
    #   torchvision
torchaudio==0.13.1
    # via -r requirements.in
torchvision==0.14.1
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   bravado
    #   swagger-spec-validator
    #   torch
    #   torchvision
tzdata==2023.3
    # via pandas
uri-template==1.2.0
    # via jsonschema
urllib3==1.26.15
    # via
    #   botocore
    #   neptune
    #   requests
webcolors==1.13
    # via jsonschema
websocket-client==1.5.1
    # via neptune
wheel==0.40.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-runtime-cu11

# The following packages are considered to be unsafe in a requirements file:
# setuptools
